Moved actor classes to "controller.actors" package
