+++ client/src/main/scala/it/cwmp/client/controller/actors/AuthenticationViewActor.scala
@@ -1,8 +1,9 @@
-package it.cwmp.client.view.authentication
+package it.cwmp.client.controller.actors
 
+import it.cwmp.client.controller.actors.AuthenticationViewActor._
 import it.cwmp.client.controller.messages.AuthenticationRequests.{LogIn, SignUp}
 import it.cwmp.client.view.FXServiceViewActor
-import it.cwmp.client.view.authentication.AuthenticationViewActor.{LOGGING_IN_MESSAGE, SIGNING_UP_MESSAGE}
+import it.cwmp.client.view.authentication.{AuthenticationFXController, AuthenticationStrategy}
 
 /**
   * Actor assigned to the management of the display of the authentication screen and of the events generated by it.
