+++ client/src/main/scala/it/cwmp/client/controller/actors/RoomViewActor.scala
@@ -1,10 +1,11 @@
-package it.cwmp.client.view.room
+package it.cwmp.client.controller.actors
 
+import it.cwmp.client.controller.actors.RoomViewActor._
 import it.cwmp.client.controller.messages.AuthenticationRequests.GUILogOut
 import it.cwmp.client.controller.messages.GUIRequest
 import it.cwmp.client.controller.messages.RoomsRequests._
 import it.cwmp.client.view.FXServiceViewActor
-import it.cwmp.client.view.room.RoomViewActor._
+import it.cwmp.client.view.room.{RoomFXController, RoomStrategy}
 
 /**
   * This class represents the actor that manages the rooms
